+++ src/main/java/webserver/request/HttpRequest.java
@@ -1,4 +1,4 @@
-package webserver.httpRequest;
+package webserver.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -7,7 +7,7 @@ import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.util.Map;
 
-import webserver.httpRequest.exception.HttpRequestParsingException;
+import webserver.request.exception.HttpRequestParsingException;
 
 public class HttpRequest {
