+++ src/main/java/webserver/request/HttpRequestHeader.java
@@ -1,4 +1,4 @@
-package webserver.httpRequest;
+package webserver.request;
 
 import java.io.BufferedReader;
 import java.io.IOException;
