+++ src/main/java/webserver/request/exception/HttpRequestParsingException.java
@@ -1,4 +1,4 @@
-package webserver.httpRequest.exception;
+package webserver.request.exception;
 
 public class HttpRequestParsingException extends RuntimeException {
