+++ src/main/java/webserver/response/HttpResponse.java
@@ -1,4 +1,4 @@
-package webserver.httpResponse;
+package webserver.response;
 
 public class HttpResponse {
